Kobo Phase 1: source abstraction, aggregator, and SQLite dependency
Add Microsoft.Data.Sqlite (with audit suppression) to the CLI, introduce the
IHighlightSource abstraction and SourceDescriptor, extract the Kindle parser's
dedup/grouping tail into a shared HighlightAggregator, and add KindleClippingsSource.
Kindle parsing behavior is unchanged.

Closes #378
