+++ specs/016-kobo/tasks.md
@@ -22,15 +22,15 @@
 
 **⚠️ CRITICAL**: No Kobo user-story work can begin until the abstraction exists and the existing Kindle parser tests still pass (the aggregator extraction is behavior-preserving).
 
-- [ ] T001 [P] Add `<MicrosoftDataSqliteVersion>10.0.5</MicrosoftDataSqliteVersion>` to the `<PropertyGroup>` in `src/PackageVersions.props` (mirrors the server's version; follows the existing `PollyVersion` convention). See `specs/016-kobo/research.md` §6.
-- [ ] T002 In `src/Relego.Cli/Relego.Cli.csproj` add `<PackageReference Include="Microsoft.Data.Sqlite" Version="$(MicrosoftDataSqliteVersion)" />` **and** an `<ItemGroup>` with `<NuGetAuditSuppress Include="https://github.com/advisories/GHSA-2m69-gcr7-jv3q" />` mirroring `Relego.Server.csproj`. Without the suppression the build fails (`TreatWarningsAsErrors=true`, GHSA-2m69-gcr7-jv3q / CVE-2025-6965). See `specs/016-kobo/research.md` §6 and `specs/016-kobo/quickstart.md`.
-- [ ] T003 [P] Create directories `src/Relego.Cli/Sources/` and `src/Relego.Tests/Sources/`, and confirm the committed fixture `docs/examples/kobo-highlights.sqlite` is present (do not regenerate it — FR-016).
-- [ ] T004 Verify `dotnet build src/Relego.slnx` succeeds with `Microsoft.Data.Sqlite` referenced and the audit suppressed. Fix any restore/build errors before proceeding.
-- [ ] T005 [P] Create the `IHighlightSource` interface and `SourceDescriptor` record in `src/Relego.Cli/Sources/IHighlightSource.cs` and `src/Relego.Cli/Sources/SourceDescriptor.cs` per the contract in `specs/016-kobo/contracts/kobo-sqlite-read.md` §2. Interface (this phase): `SourceDescriptor Descriptor { get; }` + `Task<ParseResult> ReadAsync(string path, ILogger? logger = null, CancellationToken cancellationToken = default)`. `SourceDescriptor(string Id, string DisplayName)` is the source's self-owned identity, used only as a label (no central `HighlightSourceKind` enum — keeps the registry open, ADR-008 §5). The `Locate` detection member is added in US2 (Phase 3). Namespace `Relego.Cli.Sources`.
-- [ ] T006 Create internal `HighlightAggregator` in `src/Relego.Cli/Parsing/HighlightAggregator.cs` by **extracting verbatim** the dedup + grouping tail currently inside `ClippingsParser.ParseAsync` (the text-less filter, `HashSet<(string, string?, string)>` dedup keyed on `(Title, Author, finalText)` keeping first occurrence, `(Title, Author)` first-seen grouping, empty-book exclusion, and `TotalEntriesProcessed`/`DuplicatesRemoved` counting). Signature: `ParseResult Aggregate(IReadOnlyList<RawClipping> clippings, int totalEntriesProcessed)`. Pure, no I/O. See `specs/016-kobo/data-model.md` (HighlightAggregator) and `src/Relego.Cli/Parsing/ClippingsParser.cs`.
-- [ ] T007 Refactor `src/Relego.Cli/Parsing/ClippingsParser.cs` to delegate its dedup/grouping to `HighlightAggregator.Aggregate(...)` instead of doing it inline. The `[my note]` prefix application must remain consistent so the aggregator dedups on the final prefixed text exactly as today. No change to `ClippingsParser`'s public signatures.
-- [ ] T008 [P] Create `KindleClippingsSource` in `src/Relego.Cli/Sources/KindleClippingsSource.cs`: an `IHighlightSource` with `Descriptor => new SourceDescriptor("kindle", "Kindle")` whose `ReadAsync` delegates to `ClippingsParser.ParseAsync(path, logger)`. No new parsing logic (FR-011). Its `Locate` (Kindle detection) is added in US2 (Phase 3).
-- [ ] T009 Run `dotnet test src/Relego.Tests/Relego.Tests.csproj --filter "FullyQualifiedName~Parsing"` — **all existing Kindle parser tests must still pass**, proving the aggregator extraction is behavior-preserving. Then `dotnet build src/Relego.slnx` clean.
+- [X] T001 [P] Add `<MicrosoftDataSqliteVersion>10.0.5</MicrosoftDataSqliteVersion>` to the `<PropertyGroup>` in `src/PackageVersions.props` (mirrors the server's version; follows the existing `PollyVersion` convention). See `specs/016-kobo/research.md` §6.
+- [X] T002 In `src/Relego.Cli/Relego.Cli.csproj` add `<PackageReference Include="Microsoft.Data.Sqlite" Version="$(MicrosoftDataSqliteVersion)" />` **and** an `<ItemGroup>` with `<NuGetAuditSuppress Include="https://github.com/advisories/GHSA-2m69-gcr7-jv3q" />` mirroring `Relego.Server.csproj`. Without the suppression the build fails (`TreatWarningsAsErrors=true`, GHSA-2m69-gcr7-jv3q / CVE-2025-6965). See `specs/016-kobo/research.md` §6 and `specs/016-kobo/quickstart.md`.
+- [X] T003 [P] Create directories `src/Relego.Cli/Sources/` and `src/Relego.Tests/Sources/`, and confirm the committed fixture `docs/examples/kobo-highlights.sqlite` is present (do not regenerate it — FR-016).
+- [X] T004 Verify `dotnet build src/Relego.slnx` succeeds with `Microsoft.Data.Sqlite` referenced and the audit suppressed. Fix any restore/build errors before proceeding.
+- [X] T005 [P] Create the `IHighlightSource` interface and `SourceDescriptor` record in `src/Relego.Cli/Sources/IHighlightSource.cs` and `src/Relego.Cli/Sources/SourceDescriptor.cs` per the contract in `specs/016-kobo/contracts/kobo-sqlite-read.md` §2. Interface (this phase): `SourceDescriptor Descriptor { get; }` + `Task<ParseResult> ReadAsync(string path, ILogger? logger = null, CancellationToken cancellationToken = default)`. `SourceDescriptor(string Id, string DisplayName)` is the source's self-owned identity, used only as a label (no central `HighlightSourceKind` enum — keeps the registry open, ADR-008 §5). The `Locate` detection member is added in US2 (Phase 3). Namespace `Relego.Cli.Sources`.
+- [X] T006 Create internal `HighlightAggregator` in `src/Relego.Cli/Parsing/HighlightAggregator.cs` by **extracting verbatim** the dedup + grouping tail currently inside `ClippingsParser.ParseAsync` (the text-less filter, `HashSet<(string, string?, string)>` dedup keyed on `(Title, Author, finalText)` keeping first occurrence, `(Title, Author)` first-seen grouping, empty-book exclusion, and `TotalEntriesProcessed`/`DuplicatesRemoved` counting). Signature: `ParseResult Aggregate(IReadOnlyList<RawClipping> clippings, int totalEntriesProcessed)`. Pure, no I/O. See `specs/016-kobo/data-model.md` (HighlightAggregator) and `src/Relego.Cli/Parsing/ClippingsParser.cs`.
+- [X] T007 Refactor `src/Relego.Cli/Parsing/ClippingsParser.cs` to delegate its dedup/grouping to `HighlightAggregator.Aggregate(...)` instead of doing it inline. The `[my note]` prefix application must remain consistent so the aggregator dedups on the final prefixed text exactly as today. No change to `ClippingsParser`'s public signatures.
+- [X] T008 [P] Create `KindleClippingsSource` in `src/Relego.Cli/Sources/KindleClippingsSource.cs`: an `IHighlightSource` with `Descriptor => new SourceDescriptor("kindle", "Kindle")` whose `ReadAsync` delegates to `ClippingsParser.ParseAsync(path, logger)`. No new parsing logic (FR-011). Its `Locate` (Kindle detection) is added in US2 (Phase 3).
+- [X] T009 Run `dotnet test src/Relego.Tests/Relego.Tests.csproj --filter "FullyQualifiedName~Parsing"` — **all existing Kindle parser tests must still pass**, proving the aggregator extraction is behavior-preserving. Then `dotnet build src/Relego.slnx` clean.
 
 **Checkpoint**: SQLite dependency restores, the solution builds clean, the shared abstraction and aggregator exist, and the Kindle path is unchanged in behavior. **PR-ready**: this phase ships on its own with all existing tests green. Kobo reader work can begin.
