+++ src/Relego.Tests/Sources/.gitkeep
@@ -1,0 +1,1 @@
+# Source reader tests (KoboReaderSource, HighlightSourceResolver, …) land here in later phases.
